v7a_noneon: barebox update v2024.07.0 -> v2025.03.0
Added patch for new barebox version, as a regression was introduced due
to a change in the device tree. The patch adds a new compatible string
as the old one was removed and caused problems during bootup which led to
barebox not showing any console. So user interaction was not possible.
The patch has already been applied to the upstream barebox repository in [1]

Set new option CONFIG_CONSOLE_ACTIVATE_ALL_FALLBACK to yes. As this
was found to be safest option.
All other config changes follow the default values.

[1] https://github.com/barebox/barebox/commit/9532f37b1171675e8ebfa745fbf1e2626df0a6ca