v7a: barebox: boot from bootsource
Barebox 2025.05.0 with CONFIG_BOOT_DEFAULTS (which we implicitly
enabled in the last commit) can resolve the boot target "bootsource" to
the device from which the board was initially booted, if this is
supported by the platform code (see barebox commit 6d91022db295 "common:
bootdef: add new boot entry provider" for more info). This way we won't
have to hard-code "mmc0", "mmc2", or "disk0" as the boot target for
different platforms.

Link: https://git.pengutronix.de/cgit/barebox/commit/?id=6d91022db29555c175e9
Signed-off-by: Roland Hieber <rhi@pengutronix.de>