v7a: stm32mp: switch to upstream barebox recipe
Remove the separate barebox-stm32 recipe in favour of the barebox recipe
from upstream PTXdist, which already supports multi-arch build. Merge
the kconfig options from the config diff. Let the barebox recipe install
its DTBs into the image folder (via BAREBOX_INSTALL_DTBS), which are
later picked up by genimage (however, the upstream recipe uses a
different naming scheme). The old defaultenv is already the same as in
the upstream recipe, so we don't need to port anything.

Since the barebox image is now too large to fit into the FIP partition,
enlarge the partition. (Incidentally, genimage v18 has a bug that does
not check whether the FIP image will fit into the partition, but this
will be fixed in a future release.)

Link: https://github.com/pengutronix/genimage/pull/293
Signed-off-by: Roland Hieber <rhi@pengutronix.de>