v7a: vexpress: switch to upstream barebox recipe
Remove the separate barebox-vexpress recipe in favour of the barebox
recipe from upstream PTXdist, which already supports multi-arch build.
Merge the kconfig options from the config diff. The old defaultenv is
already the same as in the upstream recipe, so we don't need to port
anything. However, the Vexpress platform cannot detect their bootsource,
so we have to introduce an init script that hard-codes our boot disk in
a suitably named boot script as a workaround so that the default boot
target can be successfully booted.

With the upstream barebox recipe, the barebox image is currently at
about 940 kiB, so we have to increase the partition sizes in the NOR
image. (Even if the env image is named "barebox-default-environment",
it is really an empty environment image only containing a correct
superblock and checksum, but no values, like calling "bareboxenv -z").

Signed-off-by: Roland Hieber <rhi@pengutronix.de>