v7a: remove outdated support for afi-gf board
With the recent changes in barebox' AM33xx platform, it will be possible
to build the MLO along with the proper barebox image, so we could get
rid of the extra barebox-am335x-mlo recipe. The upstream afi-gf board
doesn't support that new mechanism yet. However, the board also has only
historical relevance today, so remove its support in DistroKit.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>