v7a: vexpress: harmonize barebox partition layout
Since commit 5c519849029d ("v7a: vexpress: switch to upstream barebox
recipe"), we switched to the upstream barebox recipe for the vexpress
platform, which increased the size of the barebox binary to 940KiB.
However, due to the fixed partition layout in the vexpress device trees,
barebox still thinks that the environment partition starts at offset
512K, so executing a "saveenv" will overwrite the barebox image in NOR
flash, and Qemu will fail to boot on the next run.

Pick a patch from the mailing list that increases the fixed barebox
and the environment partitions in the device tree 2M each, and harmonize
our genimage setup too.

Suggested-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Roland Hieber <rhi@pengutronix.de>