ci-prepare: create r5 toolchain system independent
The previously linked toolchain is very system dependent and not
portable and might give problems down the line if the v7a compiler and
tollchain might change. Use the ptxdist with the v7a platformconfig to
get a suitable toolchain gives the same result, but much cleaner to
read.
Only downside is, that the currenlty selected toolchain would be
overwritten. Therefor move it out of the way before createing _r5
toolchain and move then back in afterwards.

Signed-off-by: Fabian Pflug <f.pflug@pengutronix.de>