platform-v8a: add nxp-imx93-frdm board
After fixing the missing upstream support in the previous two commits,
we can now fully integrate the i.MX93 FRDM board.

It recycles the image config from the TQ-MA93 board and the image file
in itself is just a copy from tqma with some name changes. Same goes for
the bootstate.dtsi. The magic value is random thoughts, the rest is copy
paste.

Signed-off-by: Fabian Pflug <f.pflug@pengutronix.de>