ptxconfig: optee: disable detached user TAs
Detached user TAs loaded from a Linux filesystem via tee-supplicant are
signed so that OP-TEE can verify their integrity and authenticity before
execution.

There are no such TAs in the BSP so we don't need the support for those,
so just disable it.

This silences our sanity checker that correctly complains that this is
not needed.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>