v7a: barebox: am335x: drop redundant usb-limit-1300 script
The script is taken from barebox and since 2014, it's being added on a
per board basis with the DistroKit provided BSP being overlaid on top.

I thus see no reason, why we this file was imported in 2018, but it's
not needed now for sure, because it just masks the equally named file
included in the original barebox built-in environment.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>